Add syntactic sugar to configure the Sampler
It is possible to initialize the `Sampler` with the model and all
parameters. To be consistent with our `using(model)` syntax, we add a
`configure` method that allows to set the value of the parameters and
return the current instance. It also allows to change the configuration
of the sampler with re-instantiating the class.